refactor(mllp-client): move node adapter to src/runtime/node.ts
Group runtime adapters under runtime/ (Node now; Deno/Workers later) rather than
at the top level. Updated the package.json ./node export
(dist/node.js → dist/runtime/node.js), the tsdown entry, the internal client
import, and the test import. The public `@glion/mllp-client/node` subpath is
unchanged. 96 tests pass; build emits dist/runtime/node.js.
